# src/main/resources/swagger/swagger.yml
openapi: 3.0.1
info:
  title: State API
  version: 1.0.0
  description: State API Swagger - Sprint 3
servers:
  - url: http://localhost:8080/
    description: localhost
paths:
  /api/states:
    post:
      tags:
        - State
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StateResponseDTO'
    get:
      tags:
        - State
      operationId: findAll
      parameters:
        - name: region
          in: query
          required: false
          schema:
            type: string
            enum:
              - NORTE
              - NORDESTE
              - SUL
              - SUDESTE
              - CENTRO_OESTE
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StateResponseParameters'
  /api/states/{id}:
    get:
      tags:
        - State
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StateResponseDTO'
    put:
      tags:
        - State
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StateResponseDTO'
    delete:
      tags:
        - State
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    StateRequestDTO:
      required:
        - area
        - capital
        - name
        - population
        - region
      type: object
      properties:
        name:
          type: string
        region:
          type: string
          enum:
            - NORTE
            - NORDESTE
            - SUL
            - SUDESTE
            - CENTRO_OESTE
        population:
          type: integer
          format: int32
        capital:
          type: string
        area:
          type: number
          format: double
    StateResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        region:
          type: string
          enum:
            - NORTE
            - NORDESTE
            - SUL
            - SUDESTE
            - CENTRO_OESTE
        population:
          type: integer
          format: int32
        capital:
          type: string
        area:
          type: number
          format: double
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    StateResponseParameters:
      type: object
      properties:
        numberOfElements:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        states:
          type: array
          items:
            $ref: '#/components/schemas/StateResponseDTO'